git-svn-id: file:///Users/raymondhs/Workspace/p/trunk@201 659f77e0-993d-4cc1-9a52-bb51f35ee826

+++  simple-spa --username raymhs91@gmail.com/SPA/PKB/ConstantTable.cpp
@@ -38,7 +38,7 @@ void DFS(TNode *root, int val, vector<TNode *> &result) {
 vector<TNode*> ConstantTable::getConstNodes(int v){
 	vector<TNode*> result;
 	vector<int> assigns = StmtTable::getStmtTable()->getAllAssign();
-	for(int i = 0; i < assigns.size(); i++) {
+	for(int i = 0; (unsigned)i < assigns.size(); i++) {
 		TNode *assign = StmtTable::getStmtTable()->getStmtNode(assigns[i]);
 		DFS(assign, v, result);
 	}
